fixing imports for tests
