+++ Task3/test/pytest.ini
@@ -1,3 +1,4 @@
+[pytest.ini]
 markers =
     input: tests for user input (command line etc.)
     reader: tests for functions dealing with reading in files
